py: Factor persistent code load/save funcs into persistentcode.[ch].
