+++ esp8266/esp8266_512k.ld
@@ -88,6 +88,7 @@ SECTIONS
         *py/builtin*.o*(.literal* .text*)
         *py/compile.o*(.literal* .text*)
         *py/emit*.o*(.literal* .text*)
+        *py/persistentcode*.o*(.literal* .text*)
         *py/formatfloat.o*(.literal* .text*)
         *py/frozenmod.o*(.literal* .text*)
         *py/gc.o*(.literal* .text*)
